Add rapidity methods (#9)
* Add rapidity methods

Add methods that return the true rapidity of LorentzVectors (E, p)
these are defined for both cartesian and cylindrical cases as rapidity()
(same method name as ScikitHEP's Vector package).

Add methods for the square of mass and the square of momentum
as these can be useful for more specialist calculations (but not
exported). 

Change mass() to be negative for vectors where p^2 > E^2, which
is the convention followed in ROOT and in ScikitHEP Vector.